+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adds support for SPL2SMAP ([SMAP L2 Radar/Radiometer Half-Orbit 9 km EASE-Grid Soil Moisture](https://nsidc.org/data/spl2smap/versions/3)).
+- Removes `tb_time_utc` variables from SPL2SMP_E output.
 
 
 ## [v0.0.4] - 2024-12-04